% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootdif.r
\name{bootdif}
\alias{bootdif}
\title{Function to compute 0.95 confidence interval for the difference in two means}
\usage{
bootdif(y, g)
}
\description{
Function to compute 0.95 confidence interval for the difference in two means
}
\note{
g is grouping variable
}
\author{
TGT
}

