% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TABLE.mktable.r
\name{mktable}
\alias{mktable}
\title{Make Table}
\usage{
mktable(x, caption = "", label = "", include.rownames = T)
}
\arguments{
\item{x}{A dataframe}

\item{caption}{The table caption}

\item{label}{The table label}

\item{include.rownames}{Whether or not the rownames should be included}
}
\description{
Convert R data frame object x to a nice pdf table via latex
}
\examples{
mktable(baseline, caption="Table 1.", label="")
}
\seealso{
xtable
}

