% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mktable2.r
\name{mktable2}
\alias{mktable2}
\title{Render multi-level table with multiple customization options enabled}
\usage{
mktable2(x, transpose = F, placement = "ht", caption = "",
  shortcap = caption, label = "", row.descrip = NULL, colnames = NULL,
  footnotes = NULL, headerlabel = NULL, size = "small",
  footnotesize = "scriptsize", labelwidth = 0.1, width = 1,
  alignment = "", use.rownames = T, use.colnames = F, shade = F,
  header = "", headerrows = 0, rnd = NULL)
}
\description{
This function provides an easy way to render a well-formated LATEX table.
}
\author{
TGT
}

