% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UTL.outvalues.r
\name{outvalues}
\alias{outvalues}
\title{Find out NAs and outliers in a variable(column)}
\usage{
outvalues(dataset, variable, iqrband=2, min=NULL, max=NULL, type=8)
}
\arguments{
\item{dataset}{Name of the dataset}

\item{variable}{Name of the variable.It should be entered as a character. For
example, "BMI".}

\item{iqrband}{The multiplier of IQR. Pick 3 for the very end extreme values.
Pick 2 to be more inclusive}

\item{min}{The smallest value that will not be determined as outliers}

\item{max}{The largest value that will not be determined as outliers}

\item{type}{An integer between 1 and 9 selecting one of the nine quantile
algorithms detailed below to be used.}
}
\value{
This function will return a new dataset with the first column as
subject/study ID and the second column as the outlying status of the input
variable. Missing values will be displayed as NA. The actual values for those
outliers will be displayed.
}
\description{
This function will find out missing values and outliers in a
certain variable in a dataset. Outliers are defined as mean +- \code{iqrband}
* interquartile range (IQR). Values exceed the \code{min} and \code{max} will
also be marked as outliers.
}
\examples{
outvalues(baseline, "BMI") # default iqrband=2
outvalues(baseline, "BMI", iqrband=3, min=15, max=35, type=8)
}

