% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UTL.sumstat.r
\name{sumstat}
\alias{sumstat}
\title{Summary Statistics}
\usage{
sumstat(a, n = F, mu = F, s = F, q = NULL, q.type = 8, mdn = F,
  mnm = F, sem = F, round.N = 3)
}
\arguments{
\item{a}{A Variable}

\item{n}{Whether the number of non-missing/null values will be displayed}

\item{mu}{Whether the value of mean will be calculated}

\item{s}{Whether the value of standard deviation will be calculated}

\item{q}{A list of quantiles}

\item{round.N}{Round digits}

\item{type}{Quantile type}
}
\description{
This function works together with data.table to generate summary
statistics information.
}
\examples{
library(data.table)
library(ifaR.handtool)
dt<-data.table(grp=rep(c("Control", "Treatment"), each=50),
               time=rep(c("Base", "Follow", "Base", "Follow"), each=25),
               age=rnorm(100, mean=65, sd=10),bmi=rnorm(100, mean=28, sd=4),
               leanmass=rnorm(100, mean=65, sd=6))
dt$leanmass[95]<-NA
dt[, as.list(unlist(lapply(.SD, sumstat, n=F, mu=T, s=T, q=c(.25, .5, .75)))), by=.(grp, time)]
}

