% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/UTL.sumtbl.r
\name{sumtbl}
\alias{sumtbl}
\title{Summary Statistics Table}
\usage{
sumtbl(dataset, variable, grp.by, n = F, mu = F, s = F, q = NULL,
  type = 8, round.N = 3)
}
\arguments{
\item{dataset}{Select the dataset you are working on}

\item{variable}{Select the variables you want to analyze}

\item{grp.by}{Select grouping variables}

\item{n}{Whether the number of non-missing/null values will be displayed}

\item{mu}{Whether the value of mean will be calculated}

\item{s}{Whether the value of standard deviation will be calculated}

\item{q}{A list of quantiles}

\item{type}{Quantile type}

\item{round.N}{Round digits}
}
\description{
This function will generate a common statistical summary table after you select
a dataset, select targeting variables, select grouping variables, select whether you want
number, mean, standard deviation and quantile.
}
\examples{
library(data.table)
library(ifaR.handtool)
dt<-data.table(grp=rep(c("Control", "Treatment"), each=50),
               time=rep(c("Base", "Follow", "Base", "Follow"), each=25),
               age=rnorm(100, mean=65, sd=10),bmi=rnorm(100, mean=28, sd=4),
               leanmass=rnorm(100, mean=65, sd=6))
dt$leanmass[95]<-NA

sumtbl(dt, c("age", "bmi"), c("grp"), n=F, mu=T, s=T, q=c(.25, .5, .75))
sumtbl(dt, c("age", "bmi", "leanmass"), c("grp", "time"), n=T, mu=T, s=T, q=c(.25, .5, .75))
}

