% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxes.r
\name{boxes}
\alias{boxes}
\title{Fancy box plots}
\usage{
boxes(y, by, col = c(rgb(0, 0, 0, 0.2), rgb(0, 0, 1, 0.2)),
  border = c("black", "blue"), medcolor = c("black", "blue"), xlab = "")
}
\description{
Fancy box plots
}
\section{Functions}{
\itemize{
\item \code{boxes}: function can produce box plots with detailed information
}}
\author{
TGT
}

