% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TIME.timetosec.r
\name{timetosec}
\alias{timetosec}
\title{mm:ss to number of seconds}
\usage{
timetosec(x)
}
\arguments{
\item{var}{Format: HH:mm:ss, mm:ss, ss}
}
\value{
A number that shows the number of seconds
}
\description{
This function converts the normal expression of time
(HH:mm:ss, mm:ss, ss) to the number of seconds. It can save a few
line of codes during walking speed calculation
}
\examples{
timetosec(sppb$time)
}

