% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TABLE.twotoone.r
\name{twotoone}
\alias{twotoone}
\title{Two to one (Table re-formating)}
\usage{
twotoone(a)
}
\arguments{
\item{a}{The input should be a summary table, whose columns are the stratifying groups
and rows are pairs of variablew in the format of "variable.Mean", "variable.SD"...etc.}
}
\description{
This function combines two elements into one in a format as "Mean (SD)".
It converts the outputs of sumstat to a simplified format, which is ready for publishing
through markdown.
}
\examples{
library(data.table)
library(ifaR.handtool)
dt<-data.table(grp=rep(c("Control", "Treatment"), each=50),
               time=rep(c("Base", "Follow", "Base", "Follow"), each=25),
               age=rnorm(100, mean=65, sd=10),bmi=rnorm(100, mean=28, sd=4),
               leanmass=rnorm(100, mean=65, sd=6))
dt$leanmass[95]<-NA

tab1<-sumtbl(dt, c("age", "bmi", "leanmass"), c("grp", "time"), n=F, mu=T, s=T, q=NULL)
twotoone(tab1)
}

