% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TABLE.mktable.r
\name{mktable}
\alias{mktable}
\title{Make Table}
\usage{
mktable(x, placement = "ht", caption = "", label = "", footnotes = NULL,
  headerlabel = "", size = "small", footnotesize = "scriptsize",
  labelwidth = 0.1, width = 1, alignment = "", use.rownames = F,
  use.colnames = F, shade = F, header = "", headercols = 0,
  labelrows = 1, rnd = NULL)
}
\arguments{
\item{x}{A dataframe}

\item{placement}{not sure what that is...}

\item{caption}{The table caption}

\item{label}{The table label}

\item{include.rownames}{Whether or not the rownames should be included}
}
\description{
Convert R data frame object x to a nice pdf table via latex
}
\examples{
mktable(baseline, caption="Table 1.", label="")
}
\seealso{
xtable
}

