% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UTL.outvalues.r
\name{outvalues}
\alias{outvalues}
\title{Find out NAs and outliers in a variable(column)}
\usage{
outvalues(dataset, variable, iqrband=2, min=NULL, max=NULL, type=8)
}
\arguments{
\item{dataset}{Name of the dataset}

\item{variable}{Name of the variable.It should be entered as a character. For
example, "BMI".}

\item{iqrband}{The IQR multiplier. Pick 3 for the very end extreme values (values > 3rd quartile + 3*IQR | values < 2nd quartile - 3*IQR) .
Pick 2 to be more inclusive.  Defaults to 2.}

\item{min}{The smallest value of the variable that will not be considered an outlier}

\item{max}{The largest value of the variable that will not be considered an outliers}

\item{type}{An integer between 1 and 9 selecting one of the nine quantile
algorithms detailed below to be used. Defaults to .}
}
\value{
This function will return a new dataset with the first column as
subject/study ID and the second column as the outlying status of the input
variable. Missing values will be displayed as NA. The actual values for those
outliers will be displayed.
}
\description{
This function will find out missing values and outliers in a
certain variable in a dataset. Outliers are defined as mean +- \code{iqrband}
* interquartile range (IQR). Values exceed the \code{min} and \code{max} will
also be marked as outliers.
}
\examples{
outvalues(baseline, "BMI") # default iqrband=2
outvalues(baseline, "BMI", iqrband=3, min=15, max=35, type=8)
}

